Move utils dir to inside the /src dir so deployment includes all necessary files.
