+++ src/steps/appliances/converter.ts
@@ -6,7 +6,7 @@ import {
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 
-import generateKey from '../../../utils/generateKey';
+import generateKey from '../../utils/generateKey';
 import { WhitehatAppliance } from '../../types';
 
 import { Entities } from '../constants';
