+++ src/steps/assets/converter.ts
@@ -7,7 +7,7 @@ import {
   parseTimePropertyValue,
 } from '@jupiterone/integration-sdk-core';
 
-import generateKey from '../../../utils/generateKey';
+import generateKey from '../../utils/generateKey';
 import { WhitehatAsset } from '../../types';
 import { Entities } from '../constants';
