+++ src/steps/findings/index.ts
@@ -4,7 +4,7 @@ import {
   IntegrationStepExecutionContext,
 } from '@jupiterone/integration-sdk-core';
 
-import generateKey from '../../../utils/generateKey';
+import generateKey from '../../utils/generateKey';
 import { createAPIClient } from '../../client';
 import { IntegrationConfig } from '../../config';
 import { WhitehatFinding } from '../../types';
