+++ src/steps/roles/converter.ts
@@ -5,7 +5,7 @@ import {
   Relationship,
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
-import generateKey from '../../../utils/generateKey';
+import generateKey from '../../utils/generateKey';
 import { WhitehatRole } from '../../types';
 
 import { Entities } from '../constants';
