+++ src/steps/roles/index.ts
@@ -4,7 +4,7 @@ import {
   IntegrationStepExecutionContext,
 } from '@jupiterone/integration-sdk-core';
 
-import generateKey from '../../../utils/generateKey';
+import generateKey from '../../utils/generateKey';
 import { createAPIClient } from '../../client';
 import { IntegrationConfig } from '../../config';
 import { WhitehatRole } from '../../types';
