+++ src/steps/scan/converter.ts
@@ -6,7 +6,7 @@ import {
   Relationship,
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
-import generateKey from '../../../utils/generateKey';
+import generateKey from '../../utils/generateKey';
 import { WhitehatAppScan, WhitehatSiteScan } from '../../types';
 
 import { Entities } from '../constants';
