+++ src/steps/users/converter.ts
@@ -6,7 +6,7 @@ import {
   Relationship,
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
-import generateKey from '../../../utils/generateKey';
+import generateKey from '../../utils/generateKey';
 import { WhitehatUser } from '../../types';
 
 import { Entities } from '../constants';
